Completes move of typing capabilities to type_utils
We have two places for type stuff -- some lived in graph.py.
Now they're both in type_utils. We also made a corresponding test file
called test_type_utils.py.
